backend/requirements.txt
fastapi
uvicorn
python-multipart

pandas
openpyxl
pdfplumber

openai
python-dotenv

psycopg2-binary
cryptography
